refactor: dashboard コンポーネントを整理する
