+++ ecscd/src/components/application/status-indicator.tsx
