+++ ecscd/src/components/dashboard/detail-pane.tsx
