+++ ecscd/src/components/dashboard/format.ts
@@ -1,0 +1,15 @@
+export function formatRelativeTime(date?: Date) {
+  if (!date) return "Never";
+
+  const now = new Date();
+  const dateObj = date instanceof Date ? date : new Date(date);
+  const diff = now.getTime() - dateObj.getTime();
+  const minutes = Math.floor(diff / 60000);
+  const hours = Math.floor(minutes / 60);
+  const days = Math.floor(hours / 24);
+
+  if (days > 0) return `${days}d ago`;
+  if (hours > 0) return `${hours}h ago`;
+  if (minutes > 0) return `${minutes}m ago`;
+  return "Just now";
+}
